requirements.txt
dualsense-controller==0.2.0
joblib==1.4.2
numpy
pyserial==3.4
rpi-lgpio==0.6
rpi-ws281x==4.2.4